[client version 5.2] - fixes bug in 5.1 (0d7ace4c567c9f9b1863701206a93ecb01d2e8c0) where user is unable to connect to server with "Error! Could not connect!"

+++ src/lihad/SYMCRelay/Client.java
@@ -104,7 +104,7 @@ public class Client implements Runnable {
 					gui.changeStatusTS(CONNECTED, true, true);
 
 					// format { <version> <username> <ip> <port> }
-					out.print(Math.floor(version) +" "+username+" "+InetAddress.getLocalHost().getHostAddress()+" "+InetAddress.getLocalHost().getHostName()+"\n"); 
+					out.print( (int)Math.floor(version) +" "+username+" "+InetAddress.getLocalHost().getHostAddress()+" "+InetAddress.getLocalHost().getHostName()+"\n"); 
 					out.flush();
